# Data-AI-Resource/BaoCao_Task21/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: CyberSoft Data & AI Lab Integration Service
  description: Cổng dịch vụ RESTful API v1.0 chuẩn hóa theo OpenAPI 3.1 cho phân hệ
    CyberSoft Data & AI Lab. Cung cấp hợp đồng tích hợp chính thức giữa phân hệ Data
    & AI Resource (TTS 01) với CyberSoft Learning & Contest Hub (TTS 02) và QA/Eval
    Platform (TTS 03).
  version: 1.0.0
paths:
...
  /api/v1/registry/datasets:
    get:
      tags:
      - Dataset Registry & Projects
      summary: Tra cứu danh sách datasets trong Registry
      description: Hỗ trợ lọc theo domain, độ khó, định dạng tệp, tag và hỗ trợ phân
        trang.
      operationId: list_datasets_api_v1_registry_datasets_get
      parameters:
      - name: domain
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Lọc theo lĩnh vực (Retail, HR/Operations, NLP/RAG, Education)
          title: Domain
        description: Lọc theo lĩnh vực (Retail, HR/Operations, NLP/RAG, Education)
      - name: difficulty_level
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Lọc theo độ khó (beginner, intermediate, advanced)
          title: Difficulty Level
        description: Lọc theo độ khó (beginner, intermediate, advanced)
      - name: format
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Lọc theo định dạng tệp (csv, json, jsonl)
          title: Format
        description: Lọc theo định dạng tệp (csv, json, jsonl)
      - name: tag
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Lọc theo thẻ phân loại
          title: Tag
        description: Lọc theo thẻ phân loại
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 50
          minimum: 1
          description: Số lượng bản ghi tối đa trả về
          default: 10
          title: Limit
        description: Số lượng bản ghi tối đa trả về
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          description: Vị trí bắt đầu
          default: 0
          title: Offset
        description: Vị trí bắt đầu
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      - name: Authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope_DatasetListResponse_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/search/semantic:
    post:
      tags:
      - Semantic Search & Retrieval
      summary: Tìm kiếm lai kết hợp ngữ nghĩa và từ khóa trên học liệu
      description: Thực thi thuật toán kết hợp BM25 và Vector Dense RRF để tìm các
        đoạn trích giáo trình phù hợp nhất.
      operationId: search_semantic_api_v1_search_semantic_post
      parameters:
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      - name: Authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SemanticSearchRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope_SemanticSearchResponse_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/tutor/chat:
    post:
      tags:
      - AI Tutor & RAG Engine
      summary: Hỏi đáp thông minh với Trợ giảng AI Tutor có trích nguồn
      description: Truy xuất kiến thức từ học liệu CyberSoft, sinh câu trả lời có
        trích dẫn minh chứng, tự động kích hoạt Guardrails và cơ chế từ chối khi câu
        hỏi ngoài phạm vi đào tạo.
      operationId: chat_with_tutor_api_v1_tutor_chat_post
      parameters:
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      - name: Authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TutorChatRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope_TutorChatResponse_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/quality/validate-dataset:
    post:
      tags:
      - Data Quality & QA Metrics
      summary: Kiểm định tính toàn vẹn và chất lượng của một dataset
      description: Thực thi các luật kiểm tra schema, missing values, duplicates và
        bounds; trả về kết luận PASSED hoặc FAILED cho Cổng chất lượng.
      operationId: validate_dataset_api_v1_quality_validate_dataset_post
      parameters:
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      - name: Authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DatasetValidationRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope_DatasetValidationResponse_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ChatMessage:
      properties:
        role:
          type: string
          title: Role
          description: 'Vai trò phát ngôn: user | assistant | system'
        content:
          type: string
          title: Content
          description: Nội dung thông điệp
      type: object
      required:
      - role
      - content
      title: ChatMessage
      description: A single turn in conversational history.
    CitationItem:
      properties:
        source_id:
          type: string
          title: Source Id
          description: Mã định danh của tài liệu trích dẫn
        title:
          type: string
          title: Title
          description: Tên giáo trình hoặc quy chế đào tạo
        section:
          type: string
          title: Section
          description: Tên mục hoặc chương cụ thể
        excerpt:
          type: string
          title: Excerpt
          description: Đoạn văn bản trích dẫn làm chứng cứ xác thực
      type: object
      required:
      - source_id
      - title
      - section
      - excerpt
      title: CitationItem
      description: Structured citation reference for verifiable factual answer.
...
    DatasetItem:
      properties:
        id:
          type: string
          title: Id
          description: Mã định danh duy nhất của dataset
        name:
          type: string
          title: Name
          description: Tên dataset
        domain:
          type: string
          title: Domain
          description: Lĩnh vực (Retail, HR/Operations, NLP/RAG, Education)
        difficulty_level:
          type: string
          title: Difficulty Level
          description: Độ khó (beginner, intermediate, advanced)
        format:
          type: string
          title: Format
          description: Định dạng tệp (csv, json, jsonl, parquet)
        records_count:
          type: integer
          title: Records Count
          description: Tổng số bản ghi dữ liệu
        current_version:
          type: string
          title: Current Version
          description: Phiên bản phát hành hiện tại
          default: v1.0
        tags:
          items:
            type: string
          type: array
          title: Tags
          description: Danh mục thẻ tìm kiếm nhanh
        is_public:
          type: boolean
          title: Is Public
          description: Học viên có quyền truy cập trực tiếp hay không
          default: true
      type: object
      required:
      - id
      - name
      - domain
      - difficulty_level
      - format
      - records_count
      title: DatasetItem
      description: Summary item of a registered dataset.
    DatasetListResponse:
      properties:
        items:
          items:
            $ref: '#/components/schemas/DatasetItem'
          type: array
          title: Items
          description: Danh sách các dataset thỏa điều kiện
        pagination:
          $ref: '#/components/schemas/PaginationMeta'
          description: Thông tin phân trang
      type: object
      required:
      - items
      - pagination
      title: DatasetListResponse
      description: Response payload for datasets list.
    DatasetValidationRequest:
      properties:
        dataset_id:
          type: string
          title: Dataset Id
          description: Mã định danh của dataset cần kiểm định
        check_rules:
          items:
            type: string
          type: array
          title: Check Rules
          description: Danh sách các bộ quy tắc kiểm định cần áp dụng
          default:
          - schema_conformance
          - missing_values
          - duplicate_rows
          - range_bounds
        max_missing_ratio_threshold:
          type: number
          maximum: 0.5
          minimum: 0.0
          title: Max Missing Ratio Threshold
          description: Ngưỡng tỷ lệ ô trống tối đa cho phép (mặc định 5%)
          default: 0.05
      type: object
      required:
      - dataset_id
      title: DatasetValidationRequest
      description: Payload to trigger an automated quality inspection on a dataset.
    DatasetValidationResponse:
      properties:
        dataset_id:
          type: string
          title: Dataset Id
          description: Mã định danh dataset được kiểm tra
        total_records:
          type: integer
          title: Total Records
          description: Tổng số dòng kiểm tra
        valid_records:
          type: integer
          title: Valid Records
          description: Số dòng dữ liệu hợp lệ
        invalid_records:
          type: integer
          title: Invalid Records
          description: Số dòng chứa sai sót hoặc vi phạm
        completeness_score:
          type: number
          title: Completeness Score
          description: Tỷ lệ đầy đủ của tập dữ liệu (0.0 đến 1.0)
        status:
          type: string
          title: Status
          description: 'Kết luận kiểm định: PASSED | WARNING | FAILED'
        passed_gate:
          type: boolean
          title: Passed Gate
          description: Đạt điều kiện cho phép xuất bản sang Learning Platform
        issues:
          items:
            $ref: '#/components/schemas/ValidationIssue'
          type: array
          title: Issues
          description: Danh sách các cảnh báo và lỗi vi phạm
      type: object
      required:
      - dataset_id
      - total_records
      - valid_records
      - invalid_records
      - completeness_score
      - status
      - passed_gate
      title: DatasetValidationResponse
      description: Inspection report detailing data quality results and gate decision.
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    PaginationMeta:
      properties:
        total:
          type: integer
          title: Total
          description: Tổng số bản ghi thỏa điều kiện
        limit:
          type: integer
          title: Limit
          description: Số bản ghi tối đa trên một trang
        offset:
          type: integer
          title: Offset
          description: Vị trí bắt đầu của danh sách kết quả
        has_next:
          type: boolean
          title: Has Next
          description: Có trang tiếp theo hay không
      type: object
      required:
      - total
      - limit
      - offset
      - has_next
      title: PaginationMeta
      description: Pagination metadata for list endpoints.
...
    SearchChunkItem:
      properties:
        chunk_id:
          type: string
          title: Chunk Id
          description: Mã định danh duy nhất của đoạn văn bản
        document_title:
          type: string
          title: Document Title
          description: Tên tài liệu / giáo trình nguồn
        section_header:
          type: string
          title: Section Header
          description: Tiêu đề mục hoặc chương sách
        content:
          type: string
          title: Content
          description: Nội dung chi tiết của đoạn trích
        relevance_score:
          type: number
          title: Relevance Score
          description: Điểm số tương đồng tổng hợp kết hợp RRF (0.0 đến 1.0)
        source_citation:
          type: string
          title: Source Citation
          description: Chuỗi trích nguồn tiêu chuẩn dạng [Doc X, Mục Y]
        tags:
          items:
            type: string
          type: array
          title: Tags
          description: Thẻ phân loại chuyên đề
      type: object
      required:
      - chunk_id
      - document_title
      - section_header
      - content
      - relevance_score
      - source_citation
      title: SearchChunkItem
      description: Retrieved chunk item with relevance score and source metadata.
    SemanticSearchRequest:
      properties:
        query:
          type: string
          maxLength: 500
          minLength: 2
          title: Query
          description: Truy vấn tìm kiếm của người dùng hoặc học viên
        top_k:
          type: integer
          maximum: 20.0
          minimum: 1.0
          title: Top K
          description: Số lượng kết quả phù hợp nhất cần trả về (1 đến 20)
          default: 5
        similarity_threshold:
          type: number
          maximum: 1.0
          minimum: 0.0
          title: Similarity Threshold
          description: Ngưỡng điểm tương đồng tối thiểu để đưa vào kết quả
          default: 0.1
        filters:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Filters
          description: Bộ lọc nâng cao (document_title, tags...)
      type: object
      required:
      - query
      title: SemanticSearchRequest
      description: Payload for executing hybrid semantic search over CyberSoft course
        corpus.
    SemanticSearchResponse:
      properties:
        query:
          type: string
          title: Query
          description: Truy vấn gốc đã thực thi
        total_found:
          type: integer
          title: Total Found
          description: Tổng số đoạn trích tìm thấy vượt qua ngưỡng lọc
        execution_time_ms:
          type: number
          title: Execution Time Ms
          description: Thời gian thực thi tìm kiếm lai tính bằng milliseconds
        results:
          items:
            $ref: '#/components/schemas/SearchChunkItem'
          type: array
          title: Results
          description: Danh sách các đoạn trích phù hợp nhất xếp theo thứ hạng
      type: object
      required:
      - query
      - total_found
      - execution_time_ms
      - results
      title: SemanticSearchResponse
      description: Response payload for semantic search.
...
    SuccessEnvelope_DatasetListResponse_:
      properties:
        success:
          type: boolean
          title: Success
          description: Trạng thái thực thi thành công luôn là true
          default: true
        data:
          $ref: '#/components/schemas/DatasetListResponse'
          description: Payload kết quả trả về của API
        meta:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Meta
          description: Metadata bổ sung (thời gian thực thi, requestId...)
      type: object
      required:
      - data
      title: SuccessEnvelope[DatasetListResponse]
    SuccessEnvelope_DatasetValidationResponse_:
      properties:
        success:
          type: boolean
          title: Success
          description: Trạng thái thực thi thành công luôn là true
          default: true
        data:
          $ref: '#/components/schemas/DatasetValidationResponse'
          description: Payload kết quả trả về của API
        meta:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Meta
          description: Metadata bổ sung (thời gian thực thi, requestId...)
      type: object
      required:
      - data
      title: SuccessEnvelope[DatasetValidationResponse]
...
    SuccessEnvelope_SemanticSearchResponse_:
      properties:
        success:
          type: boolean
          title: Success
          description: Trạng thái thực thi thành công luôn là true
          default: true
        data:
          $ref: '#/components/schemas/SemanticSearchResponse'
          description: Payload kết quả trả về của API
        meta:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Meta
          description: Metadata bổ sung (thời gian thực thi, requestId...)
      type: object
      required:
      - data
      title: SuccessEnvelope[SemanticSearchResponse]
...
    SuccessEnvelope_TutorChatResponse_:
      properties:
        success:
          type: boolean
          title: Success
          description: Trạng thái thực thi thành công luôn là true
          default: true
        data:
          $ref: '#/components/schemas/TutorChatResponse'
          description: Payload kết quả trả về của API
        meta:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Meta
          description: Metadata bổ sung (thời gian thực thi, requestId...)
      type: object
      required:
      - data
      title: SuccessEnvelope[TutorChatResponse]
...
    TutorChatRequest:
      properties:
        question:
          type: string
          maxLength: 1000
          minLength: 3
          title: Question
          description: Câu hỏi hoặc yêu cầu cần trợ giảng giải đáp
        conversation_history:
          anyOf:
          - items:
              $ref: '#/components/schemas/ChatMessage'
            type: array
          - type: 'null'
          title: Conversation History
          description: Lịch sử các lượt đối thoại trước
        top_k:
          type: integer
          maximum: 10.0
          minimum: 1.0
          title: Top K
          description: Số lượng đoạn trích giáo trình tham chiếu làm bằng chứng
          default: 3
        strict_abstention:
          type: boolean
          title: Strict Abstention
          description: Kích hoạt cơ chế từ chối nghiêm ngặt nếu câu hỏi ngoài phạm
            vi học liệu
          default: true
      type: object
      required:
      - question
      title: TutorChatRequest
      description: Payload for submitting a question to CyberSoft AI Tutor.
    TutorChatResponse:
      properties:
        answer:
          type: string
          title: Answer
          description: Nội dung giải đáp sư phạm của trợ giảng
        citations:
          items:
            $ref: '#/components/schemas/CitationItem'
          type: array
          title: Citations
          description: Danh sách các trích dẫn đối chứng trong giáo trình
        status:
          type: string
          title: Status
          description: 'Trạng thái phản hồi: ANSWERED (được trả lời) | ABSTAINED (từ
            chối do ngoài phạm vi) | GUARD_BLOCKED (chặn do bẫy bảo mật)'
        groundedness_score:
          type: number
          title: Groundedness Score
          description: Điểm số trung thực có căn cứ của câu trả lời (0.0 đến 1.0)
        latency_ms:
          type: number
          title: Latency Ms
          description: Thời gian xử lý từ lúc nhận đến hoàn thành phản hồi
      type: object
      required:
      - answer
      - status
      - groundedness_score
      - latency_ms
      title: TutorChatResponse
      description: Grounded answer from CyberSoft AI Tutor with citations and safety
        verification.
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    ValidationIssue:
      properties:
        severity:
          type: string
          title: Severity
          description: 'Mức độ nghiêm trọng: CRITICAL | WARNING | INFO'
        rule:
          type: string
          title: Rule
          description: Tên quy tắc bị vi phạm
        column:
          anyOf:
          - type: string
          - type: 'null'
          title: Column
          description: Cột dữ liệu liên quan
        description:
          type: string
          title: Description
          description: Mô tả chi tiết nguyên nhân phát sinh lỗi
      type: object
      required:
      - severity
      - rule
      - description
      title: ValidationIssue
      description: Specific defect or violation identified during data inspection.
...